% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{clean_data}
\alias{clean_data}
\title{Clean and Validate Data Frame}
\usage{
clean_data(data_frame, required_columns)
}
\arguments{
\item{data_frame}{A data frame that needs to be cleaned and validated.}

\item{required_columns}{A character vector of column names that are required in the data frame.}
}
\value{
A cleaned data frame containing only the required columns given.
}
\description{
This function checks a data frame for required columns, ensuring it's not empty,
and all required columns are present. It returns a cleaned data frame with only the required columns.
}
\examples{
required_columns <- c("shares", "num_hrefs", "num_imgs", "num_videos")
share_data <- data.frame(color = c("red","yellow","yellow"),shares = c(1,2,3), 
num_hrefs = c(1,2,3), num_imgs = c(1,2,3), num_videos = c(1,2,3),is_pop = c(0,0,0))
clean_data_frame <- clean_data(share_data, required_columns)

}
