% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_conf_mat_summary.R
\name{create_conf_mat_summary}
\alias{create_conf_mat_summary}
\title{Create and Write a Summary of the Confusion Matrix to CSV}
\usage{
create_conf_mat_summary(TP, FP, FN, TN, output_directory)
}
\arguments{
\item{TP}{True Positives count.}

\item{FP}{False Positives count.}

\item{FN}{False Negatives count.}

\item{TN}{True Negatives count.}

\item{output_directory}{Directory to save the summary CSV file.}
}
\value{
The function saves a CSV file named 'conf_mat_summary.csv'
in the specified directory and does not return anything.
}
\description{
This function calculates precision, recall, and accuracy from given
confusion matrix components (TP, FP, FN, TN) and writes the summary
into a CSV file.
}
