% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsci310utils-package.R
\docType{package}
\name{dsci310utils-package}
\alias{dsci310utils}
\alias{dsci310utils-package}
\title{dsci310utils: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
