% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histograms.R
\name{make_histogram}
\alias{make_histogram}
\title{Create a histogram with the given data}
\usage{
make_histogram(data_frame, xvar, yvar, hist_title, xlabel)
}
\arguments{
\item{data_frame}{a dataframe that need to be preprocessed}

\item{xvar}{the column/ variable for the x-axis of the histogram}

\item{yvar}{the column/ variable for the y-axis of the histogram}

\item{hist_title}{the title for the histogram}

\item{xlabel}{the label for the x-axis of the histogram}
}
\value{
a histogram with the given data and arguments
}
\description{
Create a histogram with the given data
}
