% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unzip_URL.R
\name{unzip_URL}
\alias{unzip_URL}
\title{unzip URL observations}
\usage{
unzip_URL(url, output)
}
\arguments{
\item{url}{A URL that links to a zipped folder download.}

\item{output}{Location to extract the files to}
}
\value{
Nothing.
}
\description{
Extracts the files in a zipped folder that is
downloaded from a URL.
}
\examples{
unzip_URL('https://archive.ics.uci.edu/static/public/332/online+news+popularity.zip', 'data/')
}
